feat: add pictures
