+++ src/components/ProjectCard.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import styled from 'styled-components';
+
+export const ProjectCard: React.FC = () => {
+    return (
+        <Card>
+        </Card>
+    )
+};
+
+const Card = styled.div`
+`;
